Move LanguageTable
