+++ core/src/com/unciv/ui/popups/options/LanguageTab.kt
@@ -3,7 +3,7 @@ package com.unciv.ui.popups.options
 import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.unciv.UncivGame
 import com.unciv.ui.screens.basescreen.BaseScreen
-import com.unciv.ui.components.LanguageTable.Companion.addLanguageTables
+import com.unciv.ui.components.widgets.LanguageTable.Companion.addLanguageTables
 import com.unciv.ui.components.input.onClick
 
 fun languageTab(
