+++ core/src/com/unciv/ui/screens/LanguagePickerScreen.kt
@@ -4,8 +4,8 @@ import com.unciv.Constants
 import com.unciv.models.translations.tr
 import com.unciv.ui.popups.options.OptionsPopup
 import com.unciv.ui.screens.pickerscreens.PickerScreen
-import com.unciv.ui.components.LanguageTable
-import com.unciv.ui.components.LanguageTable.Companion.addLanguageTables
+import com.unciv.ui.components.widgets.LanguageTable
+import com.unciv.ui.components.widgets.LanguageTable.Companion.addLanguageTables
 import com.unciv.ui.components.extensions.enable
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.screens.mainmenuscreen.MainMenuScreen
